update dom content
